Create ex42_playing_with_classes.py
The goal was to get user input from the command line, make an entry for a person, and then store it some place. I'm not sure set() made the most sense as a data structure for the people-book, but the thought was that it would prevent duplications, which is nice. 

I doubt this is a very pythonista way to do this, it seems more like something you'd use a dict for, but I wanted to practice making a "person" class and then storing it somewhere.