Version 2 - Modifying the dict directly
My friend and mentor Justin suggested modifying this program to do different things. The first re-working is making it so it modifies the dict directly, rather than creating an array and modifying that. I hear this difference is interacting with my data 'destructively' versus 'non-destructively'. In any case, here is another way to interact with a dict as flashcard deck.