Version 4 - Using a Deck class for the 'work'
In this version I've taking the main while loop and have taken everything out and put it into a Deck class, then call the methods of the deck class from my loop.

There was a Version 3, but in it I passed the dict each time, rather than using an __init__ function to grab and store the dict locally so it didn't have to be passed all over the place. I changed the version # locally before I realized this simply involved adding a self.terms. Oh well.