Day 02: Advent of Code solution using regex
Turns out this was a terrible idea: regexs return tuples if you have more than one group and tuples are immutable. Would have to copy my list at least twice to get from a list of tuples to a list of lists, and then potentially again to get to a list of sorted lists. This solution is terrible.