Day 03: Advent of Code solution, using list -> set
I'm unclear on if you could just use a set for the append step and trying to .add() to the set directly, or if casting the list of tuples to a set is a necessary step. One would think that the list append at each round is slower, but I don't know how fast sets are for adding or not.