Added a test case for even letters to the checker, passes.
Finished writing the main() loop and added a if __name__ == '__main__' to run it.

Also, realized that it was checking numbers and symbols for palindromes,
which means the deque worked properly, but the input wasn't a word. Added
an isaplha() test to ensure it's checking words (or at least letters).