Wrote the insert method, lots of cases. If the list is empty or if you're inserting at index 0 (front), it just calls the add method. If the position is at the end or longer than the list itself, it calls the append method. Otherwise, we know it's not the front or end, so no special cases were needed -- it just assumes an add to middle. In this case, it loops until i equals the position given, and then inserts a new node in front of the 'current' node being tracked.
Tested adding to an empty list, inserting at a position longer than the list (append),
inserting into the front of the list (add), inserting into the middle, that the order
was correctly preserved, and adding into an empty list at a position longer than the
list (should be caught by add, but if my logic is wrong, caught by append). Passes
all tests.