Got the add method working, it uses a few special cases: If the list is empty, it sets head and tail to the new node. If the new new node is smaller than head, it puts it up front and moves self.front to the new node. If the new node is larger than the tail, it puts it at the end and sets self.rear to the new node.
Otherwise, a loop traverses the list until it finds the first node greater than or
equal to the node's value, and inserts the node ahead of the current marker. I was
getting a weird logical error trying to insert after the current marker because if
current is the end of the list, then you have a None value for the .behind. Inserting
ahead of current solves this problem.

Passes test with numerical and alphabetic values.