Two versions of the same recursive problem: takes a something and reverse it.
The first uses modulo and integer division to convert a number to a string of base 2-16.
Returns the larger digits cat'd (+) the current place as a string. Moves from ones
digit up, so it looks like it adds the strings in reverse.

The second is more or less the same logic, take a string and recursively call the
function minus the first char until the string is 0 or 1, then concatenate them in
reverse order. Like a really slow, really expensive str[::-1].